refactor: extract ALE local plan CLI commands
