+++ examples/cli-agents-last-exam-command-modularization-smoke.py
@@ -37,6 +37,9 @@ def main() -> int:
     ale_source = (ROOT / "loopx" / "cli_commands" / "agents_last_exam.py").read_text(
         encoding="utf-8"
     )
+    local_plan_source = (
+        ROOT / "loopx" / "cli_commands" / "agents_last_exam_local_plan.py"
+    ).read_text(encoding="utf-8")
 
     leaked_markers = [
         "ale_local_preflight_parser = benchmark_sub.add_parser",
@@ -55,8 +58,28 @@ def main() -> int:
     assert_contains(dispatch_source, "handle_agents_last_exam_command(")
     assert_contains(init_source, "register_agents_last_exam_commands")
     assert_contains(init_source, "handle_agents_last_exam_command")
+    assert_contains(init_source, "register_agents_last_exam_local_plan_commands")
+    assert_contains(init_source, "handle_agents_last_exam_local_plan_command")
     assert_contains(ale_source, "AGENTS_LAST_EXAM_COMMANDS")
     assert_contains(ale_source, "ale-validation-run-gate")
+    assert_contains(
+        ale_source,
+        "register_agents_last_exam_local_plan_commands(",
+    )
+    assert_contains(
+        ale_source,
+        "handle_agents_last_exam_local_plan_command(",
+    )
+    for marker in (
+        "def render_agents_last_exam_local_preflight_markdown",
+        "def render_agents_last_exam_local_dry_run_plan_markdown",
+        "build_agents_last_exam_local_preflight(",
+        "build_agents_last_exam_local_dry_run_plan(",
+        'if args.benchmark_command == "ale-local-preflight":',
+    ):
+        if marker in ale_source:
+            raise AssertionError(f"{marker} leaked back into agents_last_exam.py")
+        assert_contains(local_plan_source, marker)
 
     help_result = run_cli("benchmark", "ale-validation-run-gate", "--help")
     if help_result.returncode != 0:
