+++ loopx/cli_commands/__init__.py
@@ -12,6 +12,10 @@ from .agents_last_exam import (
     handle_agents_last_exam_command,
     register_agents_last_exam_commands,
 )
+from .agents_last_exam_local_plan import (
+    handle_agents_last_exam_local_plan_command,
+    register_agents_last_exam_local_plan_commands,
+)
 from .agentissue_runner_flow import (
     handle_agentissue_runner_flow_command,
     register_agentissue_runner_flow_commands,
@@ -115,6 +119,7 @@ from .worker_bridge import handle_worker_bridge_command, register_worker_bridge_
 
 __all__ = [
     "handle_agents_last_exam_command",
+    "handle_agents_last_exam_local_plan_command",
     "handle_agentissue_runner_flow_command",
     "handle_benchmark_boundary_command",
     "handle_benchmark_command",
@@ -162,6 +167,7 @@ __all__ = [
     "handle_todo_command",
     "handle_worker_bridge_command",
     "register_agents_last_exam_commands",
+    "register_agents_last_exam_local_plan_commands",
     "register_agentissue_runner_flow_commands",
     "register_benchmark_boundary_commands",
     "register_benchmark_command_group",
